+++ src/main/java/net/ccbluex/liquidbounce/features/module/modules/movement/IceSpeed.kt
@@ -11,12 +11,17 @@ import net.ccbluex.liquidbounce.features.module.Category
 import net.ccbluex.liquidbounce.features.module.Module
 import net.ccbluex.liquidbounce.utils.block.block
 import net.ccbluex.liquidbounce.utils.extensions.isMoving
+import net.minecraft.init.Blocks.air
 import net.minecraft.init.Blocks.ice
 import net.minecraft.init.Blocks.packed_ice
 import net.minecraft.util.BlockPos
 
 object IceSpeed : Module("IceSpeed", Category.MOVEMENT) {
-    private val mode by choices("Mode", arrayOf("NCP", "AAC", "Spartan"), "NCP")
+    private val mode by choices("Mode", arrayOf("Vanilla", "NCP", "AAC", "Spartan"), "NCP")
+    private val speed by float("Speed", 0.5f, 0.2f..1.6f) { mode == "Vanilla" }
+    private val iceSlipperiness by float("IceSlipperiness", 0.98f, 0.01f..1f) { mode == "Vanilla" }
+    private val packedIceSlipperiness by float("IceSlipperiness", 0.98f, 0.01f..1f) { mode == "Vanilla" }
+
     override fun onEnable() {
         if (mode == "NCP") {
             ice.slipperiness = 0.39f
@@ -27,12 +32,20 @@ object IceSpeed : Module("IceSpeed", Category.MOVEMENT) {
 
     val onUpdate = handler<UpdateEvent> {
         val mode = mode
-        if (mode == "NCP") {
-            ice.slipperiness = 0.39f
-            packed_ice.slipperiness = 0.39f
-        } else {
-            ice.slipperiness = 0.98f
-            packed_ice.slipperiness = 0.98f
+
+        when (mode) {
+            "Vanilla" -> {
+                ice.slipperiness = iceSlipperiness
+                packed_ice.slipperiness = packedIceSlipperiness
+            }
+            "NCP" -> {
+                ice.slipperiness = 0.39f
+                packed_ice.slipperiness = 0.39f
+            }
+            else -> {
+                ice.slipperiness = 0.98f
+                packed_ice.slipperiness = 0.98f
+            }
         }
 
         val player = mc.thePlayer ?: return@handler
@@ -41,7 +54,7 @@ object IceSpeed : Module("IceSpeed", Category.MOVEMENT) {
             return@handler
         }
 
-        if (player.position.down().block.let { it != Blocks.ice && it != Blocks.packed_ice }) {
+        if (player.position.down().block.let { it != ice && it != packed_ice }) {
             return@handler
         }
 
@@ -56,7 +69,7 @@ object IceSpeed : Module("IceSpeed", Category.MOVEMENT) {
             "Spartan" -> {
                 val upBlock = BlockPos(player).up(2).block
 
-                if (upBlock != Blocks.air) {
+                if (upBlock != air) {
                     player.motionX *= 1.342
                     player.motionZ *= 1.342
                 } else {
@@ -67,12 +80,17 @@ object IceSpeed : Module("IceSpeed", Category.MOVEMENT) {
                 ice.slipperiness = 0.6f
                 packed_ice.slipperiness = 0.6f
             }
+
+            "Vanilla" -> {
+                player.motionX *= speed
+                player.motionZ *= speed
+            }
         }
     }
 
     override fun onDisable() {
-        Blocks.ice.slipperiness = 0.98f
-        Blocks.packed_ice.slipperiness = 0.98f
+        ice.slipperiness = 0.98f
+        packed_ice.slipperiness = 0.98f
         super.onDisable()
     }
 }